jwt tjek på API

+++ mongoose/controllers/matchController.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from "express";
 import Match from "../models/matchModel";
 import catchAsync from "../../utility/CatchAsync"
+import { verifyJWT } from "../../utility/Security";
+import AppError from "../../utility/AppError";
 
 export const getMatches = catchAsync( async (req: Request, res: Response) => {
 
@@ -32,25 +34,35 @@ export const getMatch = catchAsync( async (req: Request, res: Response) => {
 });
 
 export const updateMatch = catchAsync( async (req: Request, res: Response) => {
+    const token = req.headers.authorization
+    const decoded = verifyJWT(token)
+    if(decoded?.role === "Admin" || decoded?.role === "API")
+    {
 
-
+        
         const match = Match.findByIdAndUpdate(req.params.id, req.body, {
             new: true,
             runValidators: true
         })
-
-
+        
+        
         res.status(200)
-            .json({
-                status: "success",
-                match: match,
-            })
+        .json({
+            status: "success",
+            match: match,
+        })
+    }
+    throw new AppError("Not Authorized.", 401)
     
 })
 
 export const deleteMatch = catchAsync( async (req: Request, res: Response) => {
 
-    
+    const token = req.headers.authorization
+    const decoded = verifyJWT(token)
+    if(decoded?.role === "Admin" || decoded?.role === "API")
+    {
+
 
         await Match.findByIdAndDelete(req.params.id)
 
@@ -60,10 +72,16 @@ export const deleteMatch = catchAsync( async (req: Request, res: Response) => {
                 status: "success",
                 message: "Match Deleted"
             })
-    
+        }
+        throw new AppError("Not Authorized.", 401)
 })
 
 export const createMatch = catchAsync( async (req: Request, res: Response) => {
+
+    const token = req.headers.authorization
+    const decoded = verifyJWT(token)
+    if(decoded?.role === "Admin" || decoded?.role === "API")
+    {
 
         const jsonData = req.body;
         
@@ -73,5 +91,8 @@ export const createMatch = catchAsync( async (req: Request, res: Response) => {
             .json({
                 status: "success",
                 Match: newMatch
-            })    
+            })  
+        }
+    throw new AppError("Not Authorized.", 401)
+
 })
