+++ mongoose/controllers/teamController.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from "express";
 import Team from "../models/teamModel";
 import catchAsync from "../../utility/CatchAsync"
+import AppError from "../../utility/AppError";
+import { verifyJWT } from "../../utility/Security";
 
 export const getTeams = catchAsync( async (req: Request, res: Response) => {
 
@@ -32,6 +34,10 @@ export const getTeam = catchAsync( async (req: Request, res: Response) => {
 });
 
 export const updateTeam = catchAsync( async (req: Request, res: Response) => {
+    const token = req.headers.authorization
+    const decoded = verifyJWT(token)
+    if(decoded?.role === "Admin" || decoded?.role === "API")
+    {
 
 
         const team = Team.findByIdAndUpdate(req.params.id, req.body, {
@@ -45,11 +51,17 @@ export const updateTeam = catchAsync( async (req: Request, res: Response) => {
                 status: "success",
                 team: team,
             })
+
+        }
+        throw new AppError("Not Authorized.", 401)
     
 })
 
 export const deleteTeam = catchAsync( async (req: Request, res: Response) => {
-
+    const token = req.headers.authorization
+    const decoded = verifyJWT(token)
+    if(decoded?.role === "Admin" || decoded?.role === "API")
+    {
     
 
         await Team.findByIdAndDelete(req.params.id)
@@ -60,11 +72,15 @@ export const deleteTeam = catchAsync( async (req: Request, res: Response) => {
                 status: "success",
                 team: "Team Deleted"
             })
-    
+        }
+        throw new AppError("Not Authorized.", 401)
 })
 
 export const createTeam = catchAsync( async (req: Request, res: Response) => {
-
+    const token = req.headers.authorization
+    const decoded = verifyJWT(token)
+    if(decoded?.role === "Admin" || decoded?.role === "API")
+    {
         const jsonData = req.body;
         
         const newTeam = await Team.create(jsonData)
@@ -73,5 +89,7 @@ export const createTeam = catchAsync( async (req: Request, res: Response) => {
             .json({
                 status: "success",
                 Team: newTeam
-            })    
+            })  
+        }
+        throw new AppError("Not Authorized.", 401)  
 })
